requirements.txt
ccxt>=4.0.0
python-telegram-bot>=20.0
pytz>=2023.3
aiohttp>=3.9.0
python-dotenv>=1.0.0
